- Minor tidy ups from review. Highlights   - Use same url and params generation as ui thread steve requests   - Improved understandability of utils.js SteveClient     - Renamed to client file with SteveApiClient     - loadCache --> updateCache     - removed set config (which was updating) - Some WIP TODOs that i'm working through
