+++ shell/plugins/steve/getters.js
@@ -1,4 +1,4 @@
-import { isArray } from '@shell/utils/array';
+
 import { BY_TYPE } from '@shell/plugins/dashboard-store/classify';
 import { lookup } from '@shell/plugins/dashboard-store/model-loader';
 import { NAMESPACE, SCHEMA, COUNT, UI } from '@shell/config/types';
@@ -6,8 +6,9 @@ import { NAMESPACE, SCHEMA, COUNT, UI } from '@shell/config/types';
 import SteveModel from './steve-class';
 import HybridModel, { cleanHybridResources } from './hybrid-class';
 import NormanModel from './norman-class';
-import { urlFor } from '@shell/plugins/dashboard-store/getters';
 import { normalizeType } from '@shell/plugins/dashboard-store/normalize';
+import { steveUrlOptions } from './getters.utils';
+import { urlFor } from '@shell/plugins/dashboard-store/getters.utils';
 
 export const STEVE_MODEL_TYPES = {
   NORMAN:  'norman',
@@ -23,64 +24,16 @@ const GC_IGNORE_TYPES = {
 };
 
 export default {
-  urlOptions: () => (url, opt) => {
-    opt = opt || {};
-
-    // Filter
-    if ( opt.filter ) {
-      const keys = Object.keys(opt.filter);
-
-      keys.forEach((key) => {
-        let vals = opt.filter[key];
-
-        if ( !isArray(vals) ) {
-          vals = [vals];
-        }
-
-        vals.forEach((val) => {
-          url += `${ (url.includes('?') ? '&' : '?') + encodeURIComponent(key) }=${ encodeURIComponent(val) }`;
-        });
-      });
-    }
-    // End: Filter
-
-    // Limit
-    const limit = opt.limit;
-
-    if ( limit ) {
-      url += `${ url.includes('?') ? '&' : '?' }limit=${ limit }`;
-    }
-    // End: Limit
-
-    // Sort
-    const sortBy = opt.sortBy;
-
-    if ( sortBy ) {
-      url += `${ url.includes('?') ? '&' : '?' }sort=${ encodeURIComponent(sortBy) }`;
-    }
-
-    const orderBy = opt.sortOrder;
-
-    if ( orderBy ) {
-      url += `${ url.includes('?') ? '&' : '?' }order=${ encodeURIComponent(orderBy) }`;
-    }
-    // End: Sort
-
-    return url;
-  },
+  urlOptions: () => steveUrlOptions,
 
   urlFor: (state, getters) => (type, id, opt) => {
-    let url = urlFor(state, getters)(type, id, opt);
+    const url = urlFor({
+      normalizeType: getters.normalizeType, schemaFor: getters.schemaFor, baseUrl: state.config.baseUrl
+    }, {
+      type, namespace: opt.namespaced, id, opt
+    });
 
-    if (opt.namespaced) {
-      const parts = url.split('/');
-
-      url = `${ parts.join('/') }/${ opt.namespaced }`;
-
-      return url;
-    }
-
-    return url;
+    return getters.urlOptions(url, opt);
   },
 
   defaultModel: state => (obj) => {
