+++ shell/plugins/steve/mutations.js
@@ -145,7 +145,7 @@ export default {
       if (worker) {
         // Store raw json objects, not the proxies
         worker.postMessage({ loadSchemas: data });
-        // this is where I should get all the API crap to pass into the worker
+        // this is where I should get all the API crap to pass into the worker // TODO: RC check with sean
       }
     }
   },
